modularized the unit test core.
i managed to do the automatic registration for unit tests without the macro. the only thing required is inheritance from p3::unit_test::test_case.

it's the beginning of the migration process to modules. at the end, all the test groups should live in their respective modules.

structuring assertions is probably going to be a bit problematic, though. we'll see.